PS-823 add support for X-Notification-Disabled, X-Webhook-Disabled, add messenger middleware to pass quiet mode

+++ databox/api/src/Consumer/Handler/File/ImportFileHandler.php
@@ -83,7 +83,9 @@ readonly class ImportFileHandler
             $this->em->persist($file);
             $this->em->flush();
         } finally {
-            unlink($src);
+            if (file_exists($src)) {
+                unlink($src);
+            }
         }
     }
 }
